requirements.txt
openai~=1.31.2
python-dotenv~=1.0.1
selenium~=4.21.0
beautifulsoup4~=4.12.3
fastapi~=0.111.0
fastapi-responses~=0.2.1
orjson
pydantic~=2.7.3
uvicorn~=0.30.1
boto3~=1.34.81
starlette~=0.37.2
typing_extensions~=4.11.0
requests~=2.32.3